# backend/docs/swagger.yaml
basePath: /api
definitions:
  domain.AdminStats:
    properties:
      activeCoursesCount:
        type: integer
      activeUsers:
        type: integer
      newUsersLastMonth:
        type: integer
      newUsersLastWeek:
        type: integer
      storageUsed:
        description: 单位：字节
        type: integer
      systemStartTime:
        type: string
      totalCourses:
        type: integer
      totalCustomTracks:
        type: integer
      totalTracks:
        type: integer
      totalUsers:
        type: integer
    type: object
  domain.ErrorResponse:
    properties:
      code:
        type: string
      details:
        type: string
      error:
        type: string
    type: object
...
  main.ErrorResponse:
    description: API错误响应
    properties:
      error:
        description: 错误信息
        example: 未授权访问
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 提供语言学习音频播放器的后端API服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 语言学习音频播放器API
  version: "1.0"
paths:
  /api/admin/stats:
    get:
      consumes:
      - application/json
      description: 获取系统运行的统计数据
      produces:
      - application/json
      responses:
        "200":
          description: 系统统计信息
          schema:
            $ref: '#/definitions/domain.AdminStats'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取系统统计信息
      tags:
      - admin
  /api/admin/users:
    get:
      consumes:
      - application/json
      description: 获取所有系统用户的列表，支持分页
      parameters:
      - description: 页码，默认为1
        in: query
        name: page
        type: integer
      - description: 每页数量，默认为10
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 用户列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取系统用户列表
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: 创建新的系统用户
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求无效
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: 未授权访问
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建系统用户
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"